acollya/acollya-infra/requirements.txt
aws-cdk-lib==2.170.0
constructs==10.4.2
jsii>=1.106.0,<2.0.0
publication>=0.0.3
typeguard>=2.13.3,<5
